bug fix for DSCP marking
https://bugs.launchpad.net/flowgrind/+bug/1196940

Patch by Midori Kati, sent in by Lars Eggert.

+++ src/fg_socket.c
@@ -132,7 +132,7 @@ int set_window_size(int fd, int window)
 	return send < receive? send: receive;
 }
 
-int set_dscp(int fd, uint8_t dscp)
+int set_dscp(int fd, int dscp)
 {
 	int optname = IP_TOS;
 	int optlevel = IPPROTO_IP;
